refactor: Extract failure handling to separate module
Move failure handling functions to src/lib/failure-handler.js to
reduce cli.js line count from 1141 to 766 lines (under 1000 limit).

Functions moved:
- handleFailure
- detectRepository
- parseGitUrl
- isGhAuthenticated
- isGhUploadLogAvailable
- uploadLog
- canCreateIssue
- createIssue
